Add meal suggestion fields; change confidence type
Add nullable healthy_score (INT) and alternative_suggestion (NVARCHAR(500)) columns to the MEAL_SUGGESTION table in the ERD. Also change SemanticObject.Confidence from decimal? to double? in the domain entity to align the confidence value type with other numeric properties/expected data precision. Both DB columns are nullable; ensure migrations/data mappings are updated accordingly.

+++ src/SmartMarketBot.Domain/Entities/SemanticObject.cs
@@ -10,7 +10,7 @@ public class SemanticObject
     public double XMax { get; set; }
     public double YMax { get; set; }
     public string? Label { get; set; }
-    public decimal? Confidence { get; set; }
+    public double? Confidence { get; set; }
     public DateTime? DetectedAt { get; set; }
     public string? ImageUrl { get; set; }
